move audit logging test functionality into another file, so that it does not need to be pulled in for the installer
